Add support for first searching on string literals before regex
In cases like .*_PLATFORM, searching every file for that regex
is very slow. In most cases, most files don't even contain _PLATFORM.
We can use this insight to first filter on the literal string
_PLATFORM and find all cases of that, then search the whole regex
on just those matches. In a large repo, this takes a query like this
from minutes to seconds.

This can search on just a prefix, inner, or suffix string literal
to speed up the search.