Add support for -g filtering by files and directories
The -g flag can be used to include or exclude files and directories like:

-g '**/*.go' -g '!*_test.go'

This would find all matches in all .go files, but not in any of the
_test.go files.

One difference between this implementation and ripgrep's implementation
is using -g still respects .gitignore files. If a .gitignore excludes
a .go file, the above filter would include that .go file with ripgrep,
but zrep still ignores it. If it is desired to still find this file,
use the --no-ignore flag.